now handles [ss, robot demos, hand demos, same hand demos] datasets

+++ save_r3m_for_ss.py
@@ -22,6 +22,8 @@ def parse_args():
     parser.add_argument('--iou_thresh', dest='iou_thresh', type=float, required=True,
                         help='threshold for filtering data with hand and mesh bbox IoU',
                         default=0.7)
+    parser.add_argument('--robot_demos', action='store_true',
+                        help='if true, use something-something robot demo dataset')
     args = parser.parse_args()
     return args
 
@@ -39,13 +41,19 @@ def main(args):
     ## DEFINE PREPROCESSING
     transforms = T.Compose([T.Resize((224, 224)), T.ToTensor()]) # ToTensor() divides by 255
 
-    print(f'Saving R3M embeddings for data at f{args.input_dir}')
+    print(f'\nSaving R3M embeddings for {"robot" if args.robot_demos else "hand"} data at f{args.input_dir}')
     json_path = join(args.input_dir, f'IoU_{args.iou_thresh}.json')
     with open(json_path, 'r') as f:
         json_dict = json.load(f)
 
-    frames_dir = join(args.input_dir, 'frames')
-    r3m_dir = join(args.input_dir, 'r3m')
+    if args.robot_demos:
+        frames_dir = join(args.input_dir, 'robot_frames')
+        r3m_dir = join(args.input_dir, 'robot_r3m')
+    else:
+        frames_dir = join(args.input_dir, 'frames')
+        r3m_dir = join(args.input_dir, 'r3m')
+
+    n_processed, n_skipped = 0, 0
     for vid_num in tqdm(
         json_dict,
         desc=f'Going through {len(json_dict)} videos...'
@@ -55,14 +63,22 @@ def main(args):
         os.makedirs(r3m_vid_dir, exist_ok=True)
         for frame_num in json_dict[vid_num]:
             frame_img_path = join(frames_vid_dir, f'frame{frame_num}.jpg')
+            frame_r3m_path = join(r3m_vid_dir, f'frame{frame_num}_r3m.pkl')
+            if os.path.exists(frame_r3m_path):
+                n_skipped += 1
+                continue
             image = Image.open(frame_img_path)
             preprocessed_image = transforms(image).reshape(-1, 3, 224, 224)
             preprocessed_image.to(device)
             with torch.no_grad():
                 frame_r3m = r3m(preprocessed_image * 255.0)  ## R3M expects image input to be [0-255]
-            frame_r3m_path = join(r3m_vid_dir, f'frame{frame_num}_r3m.pkl')
             with open(frame_r3m_path, 'wb') as f:
                 pickle.dump(frame_r3m, f)
+            n_processed += 1
+
+    print(f'Saved R3M embeddings. Total selected frames: {n_processed + n_skipped}; '
+          f'Processed frames: {n_processed}; Skipped frames: {n_skipped}.')
+
 
 if __name__ == '__main__':
     args = parse_args()
